mspy/requirements.txt
django
django-cors-headers
imagination
pydantic
pyjwt
psycopg2-binary